adjust migration forign keys

+++ database/migrations/2020_11_08_004107_create_trips_table.php
@@ -15,7 +15,7 @@ class CreateTripsTable extends Migration
     {
         Schema::create('trips', function (Blueprint $table) {
             $table->id();
-            $table->integer('bus_id');
+            $table->unsignedBigInteger('bus_id');
             $table->dateTime('start_time');
             $table->dateTime('arrival_time');
