+++ database/migrations/2020_11_08_011429_create_seats_table.php
@@ -15,7 +15,7 @@ class CreateSeatsTable extends Migration
     {
         Schema::create('seats', function (Blueprint $table) {
             $table->id();
-            $table->integer('bus_id');
+            $table->unsignedBigInteger('bus_id');
             $table->timestamps();
 
             $table->foreign('bus_id')->references('id')->on('buses')->onDelete('cascade');
